Only remove byte versions of modules in load_module if it is older than 3 seconds to improve cuncurrency.

+++ modelmanager/utils.py
@@ -5,15 +5,18 @@ import os.path as osp
 import sys
 import fnmatch
 import shutil
+import time
 
 
 def load_module_path(path, name=None):
     """Load a python module source file python version aware."""
     name = name if name else osp.splitext(osp.basename(path))[0]
-    # remove byte versions
+    # remove byte versions if they are older than 3 seconds to avoid
+    # cuncurrency issues
     for f in [path+'c', path+'o']:
         try:
-            os.remove(f)
+            if osp.getmtime(f) - time.time() > 3:
+                os.remove(f)
         except OSError:
             pass
     # remove module references
